refactor(source): collapse factories onto root API
